requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.19
openai==1.12.0
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.4.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.6.1
streamlit==1.31.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2